+++ packages/kevin_flutter_in_app_payments/kevin_flutter_in_app_payments_ios/lib/kevin_flutter_in_app_payments_ios.dart
@@ -1,0 +1,3 @@
+library kevin_flutter_in_app_payments_ios;
+
+export 'src/kevin_flutter_payments_ios.dart';