+++ packages/kevin_flutter_in_app_payments/kevin_flutter_in_app_payments_platform_interface/lib/src/model/kevin_flutter_payments_methods.dart
@@ -1,0 +1,5 @@
+class KevinFlutterPaymentsMethods {
+  static const setPaymentsConfiguration = 'setPaymentsConfiguration';
+  static const startPayment = 'startPayment';
+  static const getCallbackUrl = 'getCallbackUrl';
+}