+++ packages/kevin_flutter_in_app_payments/kevin_flutter_in_app_payments_platform_interface/lib/src/model/kevin_session_result_payment_success.dart
@@ -6,4 +6,16 @@ class KevinSessionResultPaymentSuccess extends KevinSessionResult {
   const KevinSessionResultPaymentSuccess({
     required this.paymentId,
   });
+
+  Map<String, dynamic> toMap() {
+    return {
+      'paymentId': paymentId,
+    };
+  }
+
+  factory KevinSessionResultPaymentSuccess.fromMap(Map<String, dynamic> map) {
+    return KevinSessionResultPaymentSuccess(
+      paymentId: map['paymentId'] as String,
+    );
+  }
 }