+++ packages/kevin_flutter_in_app_payments/kevin_flutter_in_app_payments_platform_interface/lib/src/model/payment/kevin_payments_configuration.dart
@@ -1,5 +1,7 @@
+import 'package:kevin_flutter_core/kevin_flutter_core.dart';
+
 class KevinPaymentsConfiguration {
-  final String callbackUrl;
+  final KevinCallbackUrl callbackUrl;
 
   const KevinPaymentsConfiguration({
     required this.callbackUrl,