+++ packages/kevin_flutter_in_app_payments/kevin_flutter_in_app_payments_platform_interface/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.1
+
+* **BREAKING**: use KevinCallbackUrl in KevinPaymentsConfiguration instead of string
+* Upgrade kevin_flutter_core to 1.0.3
+* Add toMap/fromMap() methods to models where needed
+* Remove entity classes
+
 ## 1.0.0
 
 * Initial release.